ft(refactor): fix non-impacting bugs
