+++ src/core/actions/actionsUtils.js
@@ -1,37 +1,52 @@
 import fs from 'fs';
-import inquirer from 'inquirer';
-import pad from 'pad';
 import path from 'path';
-import {
-  removeFiles,
-  validateRemove,
-  removeRequired,
-  removeOptional,
-  createFiles,
-  createRequired,
-  createOptional,
-  selectFileToCreate,
-} from '../questions/setupQuestions/index';
-import {
-  log,
-  red,
-  gray,
-  green,
-  dimWhite,
-  useBox,
-  spinner,
-  castElementsToFormatedString,
-} from '../../common/index';
-import getInfos from '../questions/askQuestions';
-import { buildFileContent, writeFile } from '../../templates/fileWriter';
+import { fileNotDetectedAlert } from '../../common/index';
 
+/**
+ * @description
+ * gets the absolute path a file
+ *
+ * @param {String} relativePath relative path to the file
+ */
 const getResolvedPath = relativePath => path.resolve(__dirname, relativePath);
 
+/**
+ * @description
+ * Returns the activated command line option
+ *
+ * @param {Array} args all arguments available
+ * @param {String} option option activated
+ */
+const getArgs = (args, option) => args.find(item => item === option);
+
+/**
+ * @description
+ * returns an Object showing which option has been activated and the payload supplied
+ *
+ * @param {*} args all arguments available
+ * @param {*} resp option activated
+ */
+const getValues = (args, resp) => ({
+  file: getArgs(args, 'file'),
+  required: getArgs(args, 'required'),
+  optional: getArgs(args, 'optional'),
+  all: getArgs(args, 'all'),
+  isEmpty: getArgs(args, 'empty'),
+  resp,
+});
+
+/**
+ * @description
+ * Checks if file exists in the code base
+ *
+ * @param {String} relativePath relative path to file
+ */
 const checkFileExist = relativePath => {
   const resolvedPath = getResolvedPath(relativePath);
   return !!fs.existsSync(resolvedPath);
 };
 
+// required files Objects and their details
 const requiredFiles = {
   README: {
     name: 'README.md',
@@ -79,6 +94,7 @@ const requiredFiles = {
   },
 };
 
+// optional files Objects and their details
 const optionalFiles = {
   CHANGELOG: {
     name: 'CHANGELOG.md',
@@ -130,47 +146,16 @@ const optionalFiles = {
   },
 };
 
+// all files Objects and their details
 const allFiles = { ...requiredFiles, ...optionalFiles };
 
-const deleteFromCodebase = (filesArray, allFiles) => {
-  filesArray.forEach(file => {
-    const { path } = allFiles[file];
-    fs.unlink(path, err => {
-      if (err) log(`Could not delete ${file}`);
-    });
-  });
-  useBox('File(s) removed successfully \nThanks for using md-generator');
-};
+/**
+ * @description
+ * checks if file exists in the code base
 
-const deleteFiles = (ValidFilesArray, allFiles) => {
-  if (ValidFilesArray.length === 0) { return log('Error: No file selected, Please select a file\n'); }
-  inquirer
-    .prompt(validateRemove(castElementsToFormatedString(ValidFilesArray)))
-    .then(answer => {
-      const { removeFiles } = answer;
-      if (removeFiles === true) {
-        deleteFromCodebase(ValidFilesArray, allFiles);
-      }
-    });
-};
-
-const check = file => {
-  Object.keys(file).forEach(key => {
-    if (!file[key].exists) {
-      return log(
-        pad(red(' X '), 12),
-        gray(pad(`${file[key].name}`, 27), 'Not found')
-      );
-    }
-    return log(
-      pad(green(' √ '), 12),
-      dimWhite(pad(`${file[key].name}`, 28)),
-      'exists'
-    );
-  });
-  log('\n');
-};
-
+ * @param {Array} data array of files names
+ * @param {Object} allMdFiles Object containing all files available
+ */
 const checkFilesExist = (data, allMdFiles) => {
   const foundFiles = [];
   const filesNotFound = [];
@@ -188,159 +173,58 @@ const checkFilesExist = (data, allMdFiles) => {
   return { foundFiles, filesNotFound };
 };
 
+/**
+ * @description
+ * remove the .md attribute of supplied in the terminal with the file name
+ *
+ * @param {Array} item Array of file names supplied
+ */
 const removeDotMdAttribute = item => {
   const newItem = [];
   item.forEach(key => newItem.push(key.split('.')[0].toUpperCase()));
   return newItem;
 };
 
+/**
+ * @description
+ * check if file name supplied is available in the npm module
+ *
+ * @param {Array} files Array of all file names supplied
+ */
 const allExistingFiles = (files = allFiles) =>
   Object.values(files).filter(item => !!item.exists);
 
-const allFileNames = (files = allFiles) =>
+const getArrayOfValues = (files = allFiles) =>
   Object.values(files).flatMap(item => item);
 
-const processRemoval = (selectedFiles, filesInfo, mode) => {
-  if (selectedFiles.length === 0) { return log('The desired .md file(s) not found in the codebase\n'); }
-  inquirer.prompt(mode(selectedFiles)).then(answer => {
-    const selectedFiles = removeDotMdAttribute(answer.removeFiles);
-    deleteFiles(selectedFiles, filesInfo);
-  });
-};
-
-const removeNonSpecific = () => {
-  processRemoval(allExistingFiles(), allFiles, removeFiles);
-};
-
+/**
+ * @description
+ * validate files names with respect to the files supported by the npm module
+ *
+ * @param {Array} values Array of all file names supplied
+ */
 const queryFilesExistence = values => {
-  const args = values.parent.rawArgs[4];
-  if (!args || args.length === 0) {
-    return log(
-      'Error: File names not detected, please supply file names i.e --file "README.md CONTRIBUTING.md" \n'
-    );
+  const args = !values.parent ? values.join(' ') : values.parent.rawArgs[4];
+  if ((!args || args.length === 0) && values.parent) {
+    fileNotDetectedAlert();
+    process.exit(1);
   }
   const data = args.split(' ');
   return checkFilesExist(data, allFiles);
 };
-const removeSpecificFiles = values => {
-  const { foundFiles, filesNotFound } = queryFilesExistence(values);
-  if (filesNotFound.length > 0) {
-    const filesList = castElementsToFormatedString(filesNotFound);
-    log('The following file(s) were not found :\n', `${filesList} \n`);
-  }
-  if (foundFiles.length > 0) {
-    deleteFiles(foundFiles, allFiles);
-  }
-};
 
-const removeRequiredFiles = () => {
-  processRemoval(
-    allExistingFiles(requiredFiles),
-    requiredFiles,
-    removeRequired
-  );
-};
-
-const removeOptionalFiles = () => {
-  processRemoval(
-    allExistingFiles(optionalFiles),
-    optionalFiles,
-    removeOptional
-  );
-};
-
-const createMdFiles = async (USE_DEFAULT, filesToBeCreated, isEmpty) => {
-  await getInfos(USE_DEFAULT).then(projectInfos => {
-    filesToBeCreated.forEach(async file => {
-      const { path, templatePath } = file;
-      const pathToTemplate = templatePath || projectInfos.templatePath;
-      const fileContent = isEmpty
-        ? ''
-        : await buildFileContent(projectInfos, pathToTemplate);
-      writeFile(fileContent, path);
-    });
-  });
-  spinner().succeed('File(s) created Successfully\n');
-  useBox('File(s) created successfully\nThank you for using md-generator');
-};
-
-const processCreation = async (allFiles, mode, CREATE_EMPTY_FILE) => {
-  const USE_DEFAULT = true;
-  inquirer.prompt(mode(allFileNames(allFiles))).then(answer => {
-    const { createFiles } = answer;
-    const filesToBeCreated = Object.values(allFiles).filter(file =>
-      createFiles.includes(file.name));
-    if (
-      createFiles.includes('README') ||
-      createFiles.includes('CODE_OF_CONDUCT') ||
-      createFiles.includes('LICENSE')
-    ) {
-      return createMdFiles(!USE_DEFAULT, filesToBeCreated, CREATE_EMPTY_FILE);
-    }
-    createMdFiles(USE_DEFAULT, filesToBeCreated, CREATE_EMPTY_FILE);
-  });
-};
-
-const checkCreatableFiles = values => {
-  const list = typeof values[0] === 'object' ? Object.keys(values) : values;
-  const { foundFiles, filesNotFound } = queryFilesExistence(list);
-  const validFileNames = filesNotFound.filter(key =>
-    Object.keys(allFiles).includes(key));
-  const inValidFileNames = filesNotFound.filter(
-    key => !Object.keys(allFiles).includes(key)
-  );
-  if (inValidFileNames.length > 0) {
-    log(
-      `The following .md file(s) is/are not recognized as one of the required/optional .md files\n${castElementsToFormatedString(
-        inValidFileNames
-      )}`
-    );
-  }
-  if (foundFiles.length > 0) {
-    const filesList = castElementsToFormatedString(foundFiles);
-    log(
-      'The following .md file(s) already exist in the codebase :\n',
-      `${filesList} \n`
-    );
-  }
-  if (validFileNames.length > 0) {
-    const validFiles = Object.values(allFiles).filter(item =>
-      validFileNames.includes(item.name.split('.')[0]));
-    return validFiles;
-  }
-  process.exit(1);
-};
-
-const createNonSpecificFiles = () => {
-  processCreation(allFiles, selectFileToCreate);
-};
-
-const createSpecificFiles = (isEmpty, values) => {
-  const files = checkCreatableFiles(values);
-  processCreation(files, createFiles, isEmpty);
-};
-
-const createRequiredFiles = isEmpty => {
-  processCreation(requiredFiles, createRequired, isEmpty);
-};
-
-const createOptionalFiles = isEmpty => {
-  processCreation(optionalFiles, createOptional, isEmpty);
-};
+const getItemFromFileName = fileName =>
+  Object.values(allFiles).find(value => value.name === `${fileName}.md`);
 
 export {
   allFiles,
+  getValues,
   requiredFiles,
   optionalFiles,
-  check,
   checkFilesExist,
-  deleteFiles,
-  removeNonSpecific,
-  removeRequiredFiles,
-  removeOptionalFiles,
-  removeSpecificFiles,
-  createNonSpecificFiles,
-  createSpecificFiles,
-  createRequiredFiles,
-  createOptionalFiles,
+  getArrayOfValues,
+  queryFilesExistence,
+  allExistingFiles,
+  removeDotMdAttribute,
+  getItemFromFileName,
 };
