+++ src/core/questions/setupQuestions/createCommand/createInit.js
@@ -1,5 +1,14 @@
-import { castElementsToFormatedString, getFullFileNames } from '../../../../common/index';
+import {
+  castElementsToFormatedString,
+  getFullFileNames,
+} from '../../../../common/index';
 
+/**
+ * @description
+ * Question to validate files creation
+ *
+ * @param {Array} files files names array
+ */
 const createFiles = files => {
   const filesAsString = castElementsToFormatedString(getFullFileNames(files));
   return {
@@ -10,6 +19,12 @@ const createFiles = files => {
   };
 };
 
+/**
+ * @description
+ * create files Question
+ *
+ * @param {Array} files files names array
+ */
 const selectFileToCreate = files => ({
   type: 'checkbox',
   name: 'createFiles',
@@ -17,7 +32,25 @@ const selectFileToCreate = files => ({
   choices: [...getFullFileNames(files)],
 });
 
+/**
+ * @description
+ * create files Question
+ *
+ * @param {Array} files files names array
+ */
+const createEmptyFiles = () => ({
+  type: 'confirm',
+  name: 'empty',
+  message: 'Would you like the created file(s) to be empty?\n',
+  choices: false,
+});
 
+/**
+ * @description
+ * create required files Question
+ *
+ * @param {Array} files required files names array
+ */
 const createRequired = files => ({
   type: 'checkbox',
   name: 'createFiles',
@@ -26,6 +59,12 @@ const createRequired = files => ({
   choices: [...getFullFileNames(files)],
 });
 
+/**
+ * @description
+ * create optional files Question
+ *
+ * @param {Array} files optional files names array
+ */
 const createOptional = files => ({
   type: 'checkbox',
   name: 'createFiles',
@@ -34,9 +73,27 @@ const createOptional = files => ({
   choices: [...getFullFileNames(files)],
 });
 
+/**
+ * @description
+ * create optional files Question
+ *
+ * @param {Array} files optional files names array
+ */
+const overrideFiles = files => {
+  const filesAsString = castElementsToFormatedString(files);
+  return {
+    type: 'confirm',
+    name: 'override',
+    message: `The following file(s) exists,\n${filesAsString}\nwould you like to override?\n`,
+    choices: false,
+  };
+};
+
 export {
   createFiles,
   createRequired,
   createOptional,
-  selectFileToCreate
+  selectFileToCreate,
+  createEmptyFiles,
+  overrideFiles,
 };
