+++ src/core/validations/IsInvalidArgs.js
@@ -1,8 +1,24 @@
+/**
+ * @description
+ * converts array to string
+ *
+ * @param {Array} data array of information
+ */
 const str = data => JSON.stringify(data);
 
+/**
+ * @description
+ * validates  strict equality between tow items
+ * @param {Array} args
+ * @param {Array} validOptions
+ */
 const isValidOption = (args, validOptions) =>
   !validOptions.every(option => str(args) !== str(option));
 
+/**
+ * @class
+ * checks the validity of supplied arguments
+ */
 class IsValidArgs {
   static list(args) {
     const validArgs = [['required'], ['optional'], []];
